+++ better_testing/loadgen/src/token_script_burn_loadgen.ts
@@ -170,7 +170,7 @@ export async function runTokenScriptBurnLoadgen() {
   await nonceDemos.connect(bootstrapRpc)
   await nonceDemos.connectWallet(ownerMnemonic, { algorithm: "ed25519" })
   const sender = (await nonceDemos.crypto.getIdentity("ed25519")).publicKey
-  const senderHex = uint8ArrayToHex(sender)
+  const senderHex = uint8ArrayToHex(sender as Uint8Array)
   const currentNonce = await nonceDemos.getAddressNonce(senderHex)
   let nextNonce = Number(currentNonce) + 1
   const allocateNonce = () => nextNonce++
